Moves drag-and-drop demos to their own route; removes a css important directive that prevents useful things from happening
