+++ services/admin/static/style.css
@@ -116,7 +116,7 @@ h3,
 h4,
 h5,
 h6 {
-  margin: 0 0 16px 0 !important;
+  margin: 0 0 16px 0;
 }
 
 h1 {
